+++ Expanded Cheat Menu/Changelog.md
@@ -1,6 +1,8 @@
 ## ECM Changelog
 
 ## v14.6 (Unreleased)
+### Changed:
+- Examine: table/func/userdata/thread objs in tables will now use the key name for the title if there's no proper name.
 
 ## v14.5 (09 Aug 2019)
 ### Added: